Fix e2e failures: no controlLogic on frames home, repo template thumbnails
The "Update scene" feature made WorkspaceSceneDropDown and the home
dashboard scene tiles mount scenesLogic, which connects to controlLogic
and fetches /frames/{id}/states per frame on the frames home. Move
sceneUpdateVersions + updateSceneFromRepo into a new sceneUpdatesLogic
that only needs frameLogic and repositoriesModel; scenesLogic re-exposes
them for the frame workspace.

Repository templates now carry an id (their directory name), which made
Template.tsx request /templates/{id}/image — a saved-templates endpoint
that 404s for repo templates and failed the add-scene visual tests on
the frontend-error assertion. Resolve the repository image URL first and
only use templates/{id} when there is no repository.
